feat: multi-GA write-to-bus panel
Replace the single-row send bar with a 'Write to bus' panel of stacked
send rows (#215). Each row has its own GA / DPT / DPT-aware value / Write
/ Read and optional delay & cyclic scheduling, with add-row and per-row
remove controls. Immediate writes and reads work on any row.

The backend runs a single scheduled (delayed/cyclic) job at a time, so at
most one row can hold an active timer; starting another surfaces a clear
message. Removing the row that owns the running job cancels it first so no
untracked cyclic send keeps firing (per TabSel's forum note).

Supersedes SendTelegramBar; its scheduling behaviour is ported and
covered by the new panel tests.

Closes #215
