+++ frontend/src/components/WriteToBusPanel.test.tsx
@@ -1,0 +1,114 @@
+import { render, screen, fireEvent, waitFor, within } from '@testing-library/react';
+import { afterEach, beforeEach, expect, test, vi } from 'vitest';
+import { WriteToBusPanel } from './WriteToBusPanel';
+
+const IDLE = { state: 'idle' };
+const RUNNING_JOB = {
+  state: 'running', id: 'abc', address: '1/2/3', interval_seconds: 5, sends_done: 3, sends_skipped: 0,
+};
+
+function mockFetch(routes: Record<string, unknown>) {
+  return vi.fn(async (input: RequestInfo | URL) => {
+    const url = String(input);
+    for (const [path, body] of Object.entries(routes)) {
+      if (url.includes(path)) return { ok: true, json: async () => body } as Response;
+    }
+    throw new Error(`Unexpected fetch: ${url}`);
+  });
+}
+
+beforeEach(() => {
+  localStorage.clear();
+  vi.stubGlobal('fetch', mockFetch({ '/api/knx/send/scheduled/status': IDLE }));
+});
+
+afterEach(() => {
+  vi.unstubAllGlobals();
+  vi.restoreAllMocks();
+});
+
+test('sends immediately from a row when no delay or interval is set', async () => {
+  const fetchMock = mockFetch({
+    '/api/knx/send/scheduled/status': IDLE,
+    '/api/knx/send': { status: 'sent' },
+  });
+  vi.stubGlobal('fetch', fetchMock);
+
+  render(<WriteToBusPanel targets={[]} onClose={() => {}} />);
+  fireEvent.change(screen.getByPlaceholderText(/Group address/), { target: { value: '1/2/3' } });
+  fireEvent.change(screen.getByPlaceholderText(/Value/), { target: { value: '50' } });
+  fireEvent.click(screen.getByRole('button', { name: /Write/ }));
+
+  await waitFor(() => expect(screen.getByText(/Sent 50 to 1\/2\/3/)).toBeInTheDocument());
+});
+
+test('adds and removes rows, sending to multiple GAs independently (#215)', async () => {
+  const fetchMock = mockFetch({
+    '/api/knx/send/scheduled/status': IDLE,
+    '/api/knx/send': { status: 'sent' },
+  });
+  vi.stubGlobal('fetch', fetchMock);
+
+  render(<WriteToBusPanel targets={[]} onClose={() => {}} />);
+
+  // One row initially; its remove button is disabled.
+  expect(screen.getAllByPlaceholderText(/Group address/)).toHaveLength(1);
+
+  fireEvent.click(screen.getByRole('button', { name: /Add row/ }));
+  const gaInputs = screen.getAllByPlaceholderText(/Group address/);
+  expect(gaInputs).toHaveLength(2);
+
+  fireEvent.change(gaInputs[0], { target: { value: '1/2/3' } });
+  fireEvent.change(gaInputs[1], { target: { value: '4/5/6' } });
+  const valueInputs = screen.getAllByPlaceholderText(/Value/);
+  fireEvent.change(valueInputs[1], { target: { value: '1' } });
+  fireEvent.click(screen.getAllByRole('button', { name: /Write/ })[1]);
+
+  await waitFor(() => expect(screen.getByText(/Sent 1 to 4\/5\/6/)).toBeInTheDocument());
+  const sendBody = JSON.parse(
+    (fetchMock.mock.calls.find(([u]) => String(u).includes('/api/knx/send') && !String(u).includes('scheduled'))![1] as RequestInit).body as string
+  );
+  expect(sendBody.address).toBe('4/5/6');
+});
+
+test('starts a single scheduled job from a row and shows the cancel control', async () => {
+  const fetchMock = mockFetch({
+    '/api/knx/send/scheduled/status': IDLE,
+    '/api/knx/send/scheduled': RUNNING_JOB,
+  });
+  vi.stubGlobal('fetch', fetchMock);
+
+  render(<WriteToBusPanel targets={[]} onClose={() => {}} />);
+  fireEvent.change(screen.getByPlaceholderText(/Group address/), { target: { value: '1/2/3' } });
+  fireEvent.change(screen.getByPlaceholderText(/Value/), { target: { value: '50' } });
+  fireEvent.change(screen.getByPlaceholderText('Every s'), { target: { value: '5' } });
+  fireEvent.click(screen.getByRole('button', { name: /Write/ }));
+
+  await waitFor(() => expect(screen.getByText(/Cyclic send to 1\/2\/3 every 5s — 3 sent/)).toBeInTheDocument());
+  expect(screen.getByRole('button', { name: /Cancel/ })).toBeInTheDocument();
+
+  const scheduledCall = fetchMock.mock.calls.find(([u]) => String(u).endsWith('/api/knx/send/scheduled'));
+  const body = JSON.parse((scheduledCall![1] as RequestInit).body as string);
+  expect(body.interval_seconds).toBe(5);
+});
+
+test('picks up an already-running job on mount', async () => {
+  vi.stubGlobal('fetch', mockFetch({ '/api/knx/send/scheduled/status': RUNNING_JOB }));
+  render(<WriteToBusPanel targets={[]} onClose={() => {}} />);
+  await waitFor(() => expect(screen.getByText(/Cyclic send to 1\/2\/3/)).toBeInTheDocument());
+});
+
+test('DPT-1 target renders On/Off and records the GA in recents', async () => {
+  vi.stubGlobal('fetch', mockFetch({
+    '/api/knx/send/scheduled/status': IDLE,
+    '/api/knx/send': { status: 'sent' },
+  }));
+
+  render(<WriteToBusPanel targets={[{ address: '12/0/0', name: 'Heating mode', main: 1, sub: 1 }]} onClose={() => {}} />);
+  fireEvent.change(screen.getByPlaceholderText(/Group address/), { target: { value: '12/0/0' } });
+  const row = screen.getByPlaceholderText(/Group address/).closest('div')!.parentElement!;
+  fireEvent.click(within(row).getByRole('button', { name: /^On$/ }));
+
+  await waitFor(() => expect(screen.getByText(/Sent on to 12\/0\/0/)).toBeInTheDocument());
+  expect(JSON.parse(localStorage.getItem('spectrumknx-recent-send-gas')!)).toEqual(['12/0/0']);
+});
